refactor(openai): isolate Chat message mapping
Move existing normalized-message conversion out of the compatible transport implementation. This keeps the provider boundary reviewable without changing the request wire contract.
