+++ providers/openai/compatible.go
@@ -35,12 +35,6 @@ const (
 	objectModel               = "model"
 )
 
-// Content part types.
-const (
-	contentTypeImageURL = "image_url"
-	contentTypeText     = "text"
-)
-
 // Response format types.
 const (
 	responseFormatJSONObject = "json_object"
@@ -325,33 +319,6 @@ func convertAPIError(name string, apiErr *openai.Error, originalErr error) error
 	return errors.NewProviderError(name, originalErr)
 }
 
-// convertAssistantMessage converts an assistant message to OpenAI format.
-func convertAssistantMessage(msg providers.Message) openai.ChatCompletionMessageParamUnion {
-	if len(msg.ToolCalls) > 0 {
-		toolCalls := make([]openai.ChatCompletionMessageToolCallUnionParam, 0, len(msg.ToolCalls))
-		for _, tc := range msg.ToolCalls {
-			toolCalls = append(toolCalls, openai.ChatCompletionMessageToolCallUnionParam{
-				OfFunction: &openai.ChatCompletionMessageFunctionToolCallParam{
-					ID: tc.ID,
-					Function: openai.ChatCompletionMessageFunctionToolCallFunctionParam{
-						Name:      tc.Function.Name,
-						Arguments: tc.Function.Arguments,
-					},
-				},
-			})
-		}
-		return openai.ChatCompletionMessageParamUnion{
-			OfAssistant: &openai.ChatCompletionAssistantMessageParam{
-				Content: openai.ChatCompletionAssistantMessageParamContentUnion{
-					OfString: openai.String(msg.ContentString()),
-				},
-				ToolCalls: toolCalls,
-			},
-		}
-	}
-	return openai.AssistantMessage(msg.ContentString())
-}
-
 // convertChunk converts an OpenAI streaming chunk to provider format.
 func convertChunk(chunk *openai.ChatCompletionChunk) providers.ChatCompletionChunk {
 	choices := make([]providers.ChunkChoice, 0, len(chunk.Choices))
@@ -469,35 +436,6 @@ func convertEmbeddingResponse(resp *openai.CreateEmbeddingResponse) *providers.E
 	}
 
 	return result
-}
-
-// convertMessage converts a single message to OpenAI format.
-func convertMessage(msg providers.Message) (openai.ChatCompletionMessageParamUnion, error) {
-	switch msg.Role {
-	case providers.RoleAssistant:
-		return convertAssistantMessage(msg), nil
-	case providers.RoleSystem:
-		return openai.SystemMessage(msg.ContentString()), nil
-	case providers.RoleTool:
-		return openai.ToolMessage(msg.ContentString(), msg.ToolCallID), nil
-	case providers.RoleUser:
-		return convertUserMessage(msg), nil
-	default:
-		return openai.ChatCompletionMessageParamUnion{}, fmt.Errorf("unknown message role: %q", msg.Role)
-	}
-}
-
-// convertMessages converts provider messages to OpenAI format.
-func convertMessages(messages []providers.Message) ([]openai.ChatCompletionMessageParamUnion, error) {
-	result := make([]openai.ChatCompletionMessageParamUnion, 0, len(messages))
-	for _, msg := range messages {
-		converted, err := convertMessage(msg)
-		if err != nil {
-			return nil, err
-		}
-		result = append(result, converted)
-	}
-	return result, nil
 }
 
 // convertParams converts providers.CompletionParams to OpenAI request parameters.
@@ -700,27 +638,6 @@ func convertTools(tools []providers.Tool) []openai.ChatCompletionToolUnionParam 
 		})
 	}
 	return result
-}
-
-// convertUserMessage converts a user message to OpenAI format.
-func convertUserMessage(msg providers.Message) openai.ChatCompletionMessageParamUnion {
-	if msg.IsMultiModal() {
-		parts := make([]openai.ChatCompletionContentPartUnionParam, 0, len(msg.ContentParts()))
-		for _, part := range msg.ContentParts() {
-			switch part.Type {
-			case contentTypeText:
-				parts = append(parts, openai.TextContentPart(part.Text))
-			case contentTypeImageURL:
-				if part.ImageURL != nil {
-					parts = append(parts, openai.ImageContentPart(openai.ChatCompletionContentPartImageImageURLParam{
-						URL: part.ImageURL.URL,
-					}))
-				}
-			}
-		}
-		return openai.UserMessage(parts)
-	}
-	return openai.UserMessage(msg.ContentString())
 }
 
 // resolveAPIKey resolves the API key from config or environment.
